bubble chart tooltips will be html rendered
Since bubble charts tooltips aren't customizable, that was the only way I could think of to solve the duplicated field issue.

Bubble chart tooltips will be html rendered which allows you to customize its style and content in an easier way, including hiding elements. You may need to re-style your tooltip for this graph, but I don't think this is going to be a problem.

I'm not familiar with ruby on rails thats why I edited the file with pure html and css. Feel free to refuse this request and apply the idea in a proper way or just refuse the whole thing x)